feat(chrome-extension): Make sync hosts more deterministic
