+++ .changeset/wise-spiders-play.md
@@ -1,0 +1,5 @@
+---
+"@clerk/chrome-extension": major
+---
+
+Make handling of sync hosts more deterministic
