+++ packages/chrome-extension/src/internal/constants.ts
@@ -1,7 +1,4 @@
 export const CLIENT_JWT_KEY = '__client';
 export const STORAGE_KEY_CLIENT_JWT = '__clerk_client_jwt';
 export const DEFAULT_LOCAL_HOST_PERMISSION = 'http://localhost';
-export const AUTH_HEADER = {
-  production: 'Authorization',
-  development: 'Clerk-Db-Jwt',
-};
+export const AUTH_HEADER = 'Authorization';
